+++ src/shared/ui/SectionTitle/styles/SectionTitle.css
