+++ src/shared/ui/SectionTitle/ui/SectionTitle.test.tsx
